move ParallelUnitId into common, move vnode related contants into common/types
