+++ src/common/src/types/mod.rs
@@ -57,6 +57,16 @@ use crate::array::{
     StructValue,
 };
 
+/// Parallel unit is the minimal scheduling unit.
+pub type ParallelUnitId = u32;
+
+// VirtualNode (a.k.a. VNode) is a minimal partition that a set of keys belong to. It is used for
+// consistent hashing.
+pub type VirtualNode = u16;
+pub const VNODE_BITS: usize = 11;
+pub const VIRTUAL_NODE_COUNT: usize = 1 << VNODE_BITS;
+pub const VNODE_BITMAP_LEN: usize = 1 << (VNODE_BITS - 3);
+
 pub type OrderedF32 = ordered_float::OrderedFloat<f32>;
 pub type OrderedF64 = ordered_float::OrderedFloat<f64>;
