+++ src/stream/src/executor/managed_state/aggregation/extreme.rs
@@ -18,8 +18,8 @@ use madsim::collections::BTreeMap;
 use risingwave_common::array::stream_chunk::{Op, Ops};
 use risingwave_common::array::{Array, ArrayImpl};
 use risingwave_common::buffer::Bitmap;
-use risingwave_common::hash::{HashCode, VirtualNode};
-use risingwave_common::types::*;
+use risingwave_common::hash::HashCode;
+use risingwave_common::types::{VirtualNode, *};
 use risingwave_common::util::value_encoding::{deserialize_cell, serialize_cell};
 use risingwave_expr::expr::AggKind;
 use risingwave_storage::storage_value::{StorageValue, ValueMeta};
